Add files via upload
After intialising the game and the leaves (a game ending state), the minimax function is called to evaluate the leaf. 
For context:
0 = draw
-1 = o win
1 = x win 
The purpose of the copy function is to allow the algorithmn to evaluate hypothetical boards. A recursion is then used to compare the hypothetical board with the current board. Values are then exchanged, depending. on the hypothetical outcome. The "best_move" portion of the code then decides on the move to be made.